feat(train): implement multi-policy training with GPU optimization and unified checkpointing
Key changes:
- Added support for multiple cache policies: Metis (combined DQN+bandit), CACHEUS (tabular bandit), Catcher (DDPG), DQN (standalone), Bandit (contextual)
- Introduced GpuTrainingCoordinator for vectorized environment training with warmup, checkpointing and metrics
- Implemented GPU-native training with HybridRingBuffer (CPU storage, GPU sampling) to eliminate VRAM leaks
- Added unified checkpointing framework with versioning and metadata validation
- Enhanced exploration strategies: EpsilonGreedy, ThompsonSampling, UCB with GPU-native implementations
- Integrated batched action selection and shared tensor utilities for performance
- Added profiling support and comprehensive diagnostics for GPU training
- Updated configs to support warp-aligned dimensions (32) for GPU optimization
- Removed deprecated train.rs, refactored train_model.rs for multi-policy support

+++ src/models/combined.rs
@@ -155,9 +155,9 @@ impl<B: Backend> CombinedModel<B> {
             let argmax_idx = tier_q_values.argmax(1);
 
             // Convert to scalar
-            let op_idx_vec: Vec<i64> = argmax_idx
+            let op_idx_vec: Vec<i32> = argmax_idx
                 .into_data()
-                .convert::<i64>()
+                .convert::<i32>()
                 .to_vec()
                 .expect("Failed to convert argmax to vec");
             let op_idx = op_idx_vec[0];
